feat(modal): rename PaymentPlan to PaymentPlans
